Create index.html
This HTML code is for a simple web page that generates a random password. Here's a breakdown of the elements:

- <head> contains metadata about the document, including:
    - Character encoding and viewport settings
    - Links to an external stylesheet (style.css) and a shortcut icon (favicon.png)
    - A script tag that links to an external JavaScript file (main.js)
    - The title of the page ("Random Password Generator")
- <body> contains the page content, including:
    - A heading (<h2>) with the title "Random Password Generator"
    - A div with an input field (readonly) to display the generated password
    - A div with an ID of "btn" that acts as a button to generate the password
    - A footer with a paragraph containing:
        - A heart symbol (❤️)
        - A link to the author's website (Swapnil Srivastava)
        - A link to the GitHub repository for this project (#30DaysOfJavaScript)

The JavaScript code (in main.js) is likely responsible for generating the random password and updating the input field when the button is clicked.

Note: The HTML entities in the footer (&#x3c;, &#47;, &#x3e;) represent the less-than, forward slash, and greater-than symbols, respectively.